Renamed /black command to /blacklist
Previously, I was testing the command on a server that already had a plugin using the /blacklist command. To avoid conflicts during testing, I temporarily changed the command to /black. Now that the conflict is no longer an issue, I'm reverting it to /blacklist to better reflect its purpose and ensure consistency.